moved navigation to comp
