+++ src/components/Navigation.js
@@ -1,9 +1,9 @@
 import React, {useState} from 'react';
-import './CSS/App.css';
+import '../CSS/App.css';
 import { Switch, Route, Link, Redirect, NavLink, useLocation } from 'react-router-dom';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import { faHome, faUsers } from '@fortawesome/free-solid-svg-icons';
-import {LogOutButton} from './components/Buttons.js';
+import {LogOutButton} from './Buttons.js';
 
 
 export function NavBar(){
